Update to pricing with new reference price for ring0
