+++ models/credmark/protocols/dexes/uniswap/uniswap_v2.py
@@ -1,6 +1,5 @@
-# pylint: disable=too-many-lines, unsubscriptable-object
-import math
-from typing import List
+# pylint: disable=too-many-lines, unsubscriptable-object, line-too-long
+from typing import List, Tuple
 
 import numpy as np
 import pandas as pd
@@ -8,14 +7,13 @@ from credmark.cmf.model import Model
 from credmark.cmf.model.errors import ModelDataError, ModelRunError
 from credmark.cmf.types import (Address, Contract, Contracts,
                                 Maybe, Portfolio, Position, Price,
-                                PriceWithQuote, Some, Token, Tokens,
-                                Records)
+                                PriceWithQuote, Some, Token, Tokens)
 from credmark.cmf.types.block_number import BlockNumberOutOfRangeError
 from credmark.cmf.types.compose import MapInputsOutput
-from credmark.dto import DTO, EmptyInput, DTOField
+from credmark.dto import DTO, EmptyInput
+
 from models.credmark.price.dex import get_primary_token_tuples
 from models.credmark.protocols.dexes.uniswap.constant import V2_FACTORY_ADDRESS
-from models.credmark.protocols.dexes.uniswap.types import PositionWithFee
 from models.dtos.price import (DexPricePoolInput, DexPriceTokenInput)
 from models.dtos.pool import PoolPriceInfo
 from models.dtos.tvl import TVLInfo
@@ -25,9 +23,8 @@ from web3.exceptions import ABIFunctionNotFound, BadFunctionCallOutput, Contract
 
 class UniswapV2PoolMeta:
     @staticmethod
-    def get_uniswap_pools(context, input_address: Address, factory_addr: Address) -> Contracts:
+    def get_uniswap_pools_by_pair(_context, factory_addr: Address, token_pairs) -> Contracts:
         factory = Contract(address=factory_addr)
-        token_pairs = get_primary_token_tuples(context, input_address)
 
         contracts = []
         try:
@@ -50,10 +47,16 @@ class UniswapV2PoolMeta:
             return Contracts(contracts=[])
 
     @staticmethod
-    def get_uniswap_pools_ledger(context, input_address: Address, _gw: Contract) -> Contracts:
+    def get_uniswap_pools(_context, factory_addr: Address, input_address: Address) -> Contracts:
+        token_pairs = get_primary_token_tuples(_context, input_address)
+        return UniswapV2PoolMeta.get_uniswap_pools_by_pair(_context, factory_addr, token_pairs)
+
+    @staticmethod
+    def get_uniswap_pools_ledger(context, factory_addr: Address, input_address: Address) -> Contracts:
+        factory = Contract(address=factory_addr)
         token_pairs = get_primary_token_tuples(context, input_address)
 
-        with _gw.ledger.events.PairCreated as q:
+        with factory.ledger.events.PairCreated as q:
             tp = token_pairs[0]
             eq_conds = q.EVT_TOKEN0.eq(tp[0]).and_(q.EVT_TOKEN1.eq(tp[1])).parentheses_()
             for tp in token_pairs[1:]:
@@ -81,6 +84,108 @@ class UniswapV2PoolMeta:
 
         return Contracts(contracts=[Contract(c) for c in evt_pair])
 
+    WEIGHT_POWER = 4
+
+    @staticmethod
+    def get_uniswap_pool_info(_context, pool_addr: Address) -> Tuple[float, float, float, float]:
+        pool = Contract(address=pool_addr)
+        pool.set_abi(abi=UNISWAP_V2_POOL_ABI, set_loaded=True)
+
+        reserves = pool.functions.getReserves().call()
+
+        if reserves == [0, 0, 0]:
+            return 0, 0, 0, 0
+
+        token0_addr = pool.functions.token0().call()
+        token1_addr = pool.functions.token1().call()
+
+        try:
+            token0 = Token(address=Address(token0_addr)).as_erc20(set_loaded=True)
+        except (OverflowError, ContractLogicError):
+            token0 = Token(address=Address(token0_addr)).as_erc20()
+
+        try:
+            token1 = Token(address=Address(token1_addr)).as_erc20(set_loaded=True)
+        except (OverflowError, ContractLogicError):
+            token1 = Token(address=Address(token1_addr)).as_erc20()
+
+        scaled_reserve0 = token0.scaled(reserves[0])
+        scaled_reserve1 = token1.scaled(reserves[1])
+
+        try:
+            tick_price0 = scaled_reserve1 / scaled_reserve0
+            tick_price1 = 1 / tick_price0
+        except (FloatingPointError, ZeroDivisionError):
+            tick_price0 = 0
+            tick_price1 = 0
+
+        full_tick_liquidity0 = scaled_reserve0
+        one_tick_liquidity0 = np.abs(1 / np.sqrt(1 + 0.0001) - 1) * full_tick_liquidity0
+
+        full_tick_liquidity1 = scaled_reserve1
+        one_tick_liquidity1 = (np.sqrt(1 + 0.0001) - 1) * full_tick_liquidity1
+
+        return (tick_price0, one_tick_liquidity0, tick_price1, one_tick_liquidity1)
+
+    @staticmethod
+    def get_ref_price(_context, factory_addr: Address):
+        ring0_tokens = _context.run_model('dex.ring0-tokens',
+                                          input=EmptyInput(),
+                                          return_type=Some[Address],
+                                          local=True).some
+
+        ratios = {}
+        for token0_address in ring0_tokens:
+            for token1_address in ring0_tokens:
+                # Uniswap builds pools with token0 < token1
+                if token0_address.to_int() <= token1_address.to_int():
+                    continue
+                token_pairs = [(token1_address, token0_address)]
+                pools = UniswapV2PoolMeta.get_uniswap_pools_by_pair(_context, factory_addr, token_pairs)
+
+                # print((token1_address, token2_address, len(pools.contracts), pools))
+                pool_info = pd.DataFrame(
+                    data=[UniswapV2PoolMeta.get_uniswap_pool_info(_context, c.address) for c in pools.contracts],
+                    columns=['tick_price0', 'one_tick_liquidity0',
+                             'tick_price1', 'one_tick_liquidity1'])
+
+                if pool_info.shape[0] > 1:
+                    ratio0 = (pool_info.tick_price0 * pool_info.one_tick_liquidity0 ** UniswapV2PoolMeta.WEIGHT_POWER).sum() / \
+                        (pool_info.one_tick_liquidity0 ** UniswapV2PoolMeta.WEIGHT_POWER).sum()
+                    ratio1 = (pool_info.tick_price1 * pool_info.one_tick_liquidity1 ** UniswapV2PoolMeta.WEIGHT_POWER).sum() / \
+                        (pool_info.one_tick_liquidity1 ** UniswapV2PoolMeta.WEIGHT_POWER).sum()
+                else:
+                    ratio0 = pool_info['tick_price0'][0]
+                    ratio1 = pool_info['tick_price1'][0]
+
+                ratios[(token1_address, token0_address)] = ratio1
+                ratios[(token0_address, token1_address)] = ratio0
+
+        candidate_prices = []
+        price_span = np.array([])
+        for pivot_token in ring0_tokens:
+            candidate_price = np.array([ratios[(pivot_token, token)] if token !=
+                                        pivot_token else 1 for token in ring0_tokens])
+            price_span = np.append(price_span, candidate_price.max() / candidate_price.min())
+            candidate_prices.append(candidate_price / candidate_price.max())
+
+        return dict(zip(ring0_tokens,
+                    candidate_prices[np.where(price_span == price_span.min())[0][0]]))
+
+
+@Model.describe(slug='uniswap-v2.get-ring0-ref-price',
+                version='0.1',
+                display_name='Uniswap v2 Ring0 Reference Price',
+                description='The Uniswap v2 pools that support the ring0 tokens',
+                category='protocol',
+                subcategory='uniswap-v2',
+                input=EmptyInput,
+                output=dict)
+class UniswapV2GetRing0RefPrice(Model, UniswapV2PoolMeta):
+    def run(self, input: EmptyInput) -> dict:
+        factory_addr = V2_FACTORY_ADDRESS[self.context.network]
+        return self.get_ref_price(self.context, factory_addr)
+
 
 @Model.describe(slug='uniswap-v2.get-pools',
                 version='1.8',
@@ -92,8 +197,8 @@ class UniswapV2PoolMeta:
                 output=Contracts)
 class UniswapV2GetPoolsForToken(Model, UniswapV2PoolMeta):
     def run(self, input: Token) -> Contracts:
-        addr = V2_FACTORY_ADDRESS[self.context.network]
-        return self.get_uniswap_pools(self.context, input.address, Address(addr))
+        factory_addr = V2_FACTORY_ADDRESS[self.context.network]
+        return self.get_uniswap_pools(self.context, factory_addr, input.address)
 
 
 @Model.describe(slug='uniswap-v2.get-pools-ledger',
@@ -106,409 +211,12 @@ class UniswapV2GetPoolsForToken(Model, UniswapV2PoolMeta):
                 output=Contracts)
 class UniswapV2GetPoolsForTokenLedger(Model, UniswapV2PoolMeta):
     def run(self, input: Token) -> Contracts:
-        gw = Contract(V2_FACTORY_ADDRESS[self.context.network])
-        return self.get_uniswap_pools_ledger(self.context, input.address, gw)
-
-
-class V2LPInput(DTO):
-    pool: Token
-    lp: Address = DTOField(description='Account')
-
-
-class V2LPOutput(DTO):
-    lp: Position
-    tokens: List[Position]
-
-
-class V2LPFeeOutput(DTO):
-    lp: Position
-    tokens: List[PositionWithFee]
-
-    @classmethod
-    # pylint: disable=invalid-name
-    def zero(cls, lp, token0, token1):
-        lp_pos = Position(amount=0, asset=lp)
-        token0_pos = PositionWithFee(amount=0, fee=0, asset=token0)
-        token1_pos = PositionWithFee(amount=0, fee=0, asset=token1)
-        return cls(lp=lp_pos,
-                   tokens=[token0_pos, token1_pos])
-
-
-class V2LPQuantityInput(DTO):
-    pool: Token
-    lp_balance: float
-
-
-@Model.describe(slug='uniswap-v2.lp-pos',
-                version='0.2',
-                display_name='Uniswap v2 (Sushiswap) LP Position (inclusive of fee) for liquidity',
-                description='Returns position (inclusive of fee) for the amount of liquidity',
-                category='protocol',
-                subcategory='uniswap-v2',
-                input=V2LPQuantityInput,
-                output=V2LPOutput)
-class UniswapV2LPQuantity(Model):
-    def run(self, input: V2LPQuantityInput) -> V2LPOutput:
-        pool = input.pool
-        lp_balance = input.lp_balance
-
-        try:
-            _ = pool.abi
-        except ModelDataError:
-            pool.set_abi(UNISWAP_V2_POOL_ABI)
-
-        reserves = pool.functions.getReserves().call()
-        lp_total_supply = pool.functions.totalSupply().call()
-
-        token0_addr = pool.functions.token0().call()
-        token1_addr = pool.functions.token1().call()
-
-        try:
-            token0 = Token(address=Address(token0_addr)).as_erc20(set_loaded=True)
-            scaled_reserve0 = token0.scaled(reserves[0])
-        except OverflowError:
-            token0 = Token(address=Address(token0_addr)).as_erc20()
-            scaled_reserve0 = token0.scaled(reserves[0])
-
-        try:
-            token1 = Token(address=Address(token1_addr)).as_erc20(set_loaded=True)
-            scaled_reserve1 = token1.scaled(reserves[1])
-        except OverflowError:
-            token1 = Token(address=Address(token1_addr)).as_erc20()
-            scaled_reserve1 = token1.scaled(reserves[1])
-
-        if math.isclose(lp_balance, 0):
-            return V2LPOutput(
-                lp=Position(amount=lp_balance, asset=pool),
-                tokens=[Position(amount=0, asset=token0), Position(amount=0, asset=token1)])
-
-        lp_token0 = scaled_reserve0 * lp_balance / lp_total_supply
-        lp_token1 = scaled_reserve1 * lp_balance / lp_total_supply
-
-        lp_position = Position(amount=pool.scaled(lp_balance), asset=pool)
-        position0 = Position(amount=lp_token0, asset=token0)
-        position1 = Position(amount=lp_token1, asset=token1)
-
-        out = V2LPOutput(lp=lp_position, tokens=[position0, position1])
-        return out
-
-
-@Model.describe(slug='uniswap-v2.lp',
-                version='0.3',
-                display_name='Uniswap v2 (Sushiswap) LP Position (inclusive of fee) for account',
-                description='Returns position (inclusive of fee) for account',
-                category='protocol',
-                subcategory='uniswap-v2',
-                input=V2LPInput,
-                output=V2LPOutput)
-class UniswapV2LP(Model):
-    def run(self, input: V2LPInput) -> V2LPOutput:
-        pool = input.pool
-        lp = input.lp
-        lp_balance = pool.functions.balanceOf(lp).call()
-
-        return self.context.run_model(
-            'uniswap-v2.lp-pos',
-            input=V2LPQuantityInput(pool=input.pool, lp_balance=lp_balance),
-            return_type=V2LPOutput)
-
-
-# pylint: disable=invalid-name
-def calculate_v2_fee(context, pool, lp, block_number, transaction_value,
-                     lp_prev_token0, lp_prev_token1):
-
-    # Get the amount of tokens for a given amount of LP tokens.
-    lp_in_out = context.run_model(
-        'uniswap-v2.lp-pos',
-        input=V2LPQuantityInput(pool=pool, lp_balance=1e18),
-        return_type=V2LPOutput,
-        block_number=block_number)
-
-    ratio = lp_in_out.tokens[1].amount / lp_in_out.tokens[0].amount
-
-    # Position implied from previous LP position (without fee)
-    token0_lp_current = (lp_prev_token0 * lp_prev_token1 / ratio) ** 0.5
-    token1_lp_current = token0_lp_current * ratio
-
-    # LP position at block_number from LP token (with fee)
-    # $x = lp / lp_total_supply * token_x$
-    # $y = lp / lp_total_supply * token_y$
-    default_block = context.web3.eth.default_block
-    context.web3.eth.default_block = block_number
-    lp_balance = pool.functions.balanceOf(lp).call()
-    context.web3.eth.default_block = default_block
-
-    if lp_balance != 0:
-        lp_pos_token0 = lp_in_out.tokens[0].amount * lp_balance / 1e18
-        lp_pos_token1 = lp_in_out.tokens[1].amount * lp_balance / 1e18
-    else:
-        lp_pos_token0, lp_pos_token1 = 0.0, 0.0
-
-    # LP position from recent deposit/withdraw (no contribution to fee)
-    if transaction_value != 0:
-        in_out_amount0 = lp_in_out.tokens[0].amount * transaction_value / 1e18
-        in_out_amount1 = lp_in_out.tokens[1].amount * transaction_value / 1e18
-    else:
-        in_out_amount0, in_out_amount1 = 0.0, 0.0
-
-    # fee = With fee - Without fee - Just-in
-    # 1. "With fee" uses end-of-block lp token holding to calculate - up-to-date
-    # 2. "Without fee" uses previous end-of-block holding to calculate - up-to-date
-    # 3. "Just in" is calculated from the current end-of-block's ratio,
-    #    which may not be the original amount put in due to other swaps inside the block.
-    return {
-        'token0_lp': lp_pos_token0,
-        'token1_lp':  lp_pos_token1,
-        'in_out_amount0': try_zero(in_out_amount0),
-        'in_out_amount1': try_zero(in_out_amount1),
-        'token0_lp_current':  token0_lp_current,
-        'token1_lp_current': token1_lp_current,
-        'token0_fee': try_zero(lp_pos_token0 - token0_lp_current - in_out_amount0),
-        'token1_fee': try_zero(lp_pos_token1 - token1_lp_current - in_out_amount1),
-    }
-
-
-def try_zero(flt):
-    if math.isclose(flt, 0):
-        return 0
-    return flt
-
-
-def uniswap_v2_fee_sample_data():
-    _df_sample = pd.DataFrame(
-        columns=['block_number', 'log_index', 'from_address', 'to_address', 'transaction_value'],
-        data=[(int(10109485), int(173),
-              '0x0000000000000000000000000000000000000000',
-               '0x76e2e2d4d655b83545d4c50d9521f5bc63bc5329',
-               int(1885295466071170)),
-              (int(10431987), int(27),
-              '0x76e2e2d4d655b83545d4c50d9521f5bc63bc5329',
-               '0x7fba4b8dc5e7616e59622806932dbea72537a56b',
-               int(-357094337512011)),
-              (int(10933740), int(178),
-              '0x0000000000000000000000000000000000000000',
-               '0x76e2e2d4d655b83545d4c50d9521f5bc63bc5329',
-               int(75011156151825)), ])
-
-    return _df_sample
-
-
-# pylint: disable=line-too-long
-@Model.describe(slug='uniswap-v2.lp-fee-history',
-                version='1.0',
-                display_name='Uniswap v2 (Sushiswap) LP Position and Fee history for account',
-                description='Returns LP Position and Fee history for account',
-                category='protocol',
-                subcategory='uniswap-v2',
-                input=V2LPInput,
-                output=Records)
-class UniswapV2LPFeeHistory(Model):
-    def run(self, input: V2LPInput) -> Records:
-        pool = input.pool
-        lp = input.lp
-
-        try:
-            _ = pool.abi
-        except ModelDataError:
-            pool.set_abi(UNISWAP_V2_POOL_ABI)
-
-        token0 = Token(address=Address(pool.functions.token0().call()))
-        token1 = Token(address=Address(pool.functions.token1().call()))
-        token0 = token0.as_erc20(set_loaded=True)
-        token1 = token1.as_erc20(set_loaded=True)
-
-        with self.context.ledger.TokenBalance as q:
-            q_cols = [q.TRANSACTION_HASH,
-                      q.BLOCK_NUMBER,
-                      q.LOG_INDEX,
-                      q.FROM_ADDRESS,
-                      q.TO_ADDRESS,
-                      q.TRANSACTION_VALUE]
-
-        def _use_ledger():
-            with self.context.ledger.TokenBalance as q:
-                df_ts = []
-                offset = 0
-
-                q_cols = [q.BLOCK_NUMBER,
-                          q.LOG_INDEX,
-                          q.FROM_ADDRESS,
-                          q.TO_ADDRESS,
-                          q.TRANSACTION_VALUE]
-                while True:
-                    df_tt = q.select(
-                        columns=q_cols,
-                        order_by=q.BLOCK_NUMBER,
-                        where=(q.ADDRESS.eq(lp)
-                               .and_(q.TOKEN_ADDRESS.eq(pool.address))
-                               )
-                    ).to_dataframe()
-
-                    if df_tt.shape[0] > 0:
-                        df_ts.append(df_tt)
-                    if df_tt.shape[0] < 5000:
-                        break
-                    offset += 5000
-
-            _df = pd.DataFrame()
-            if len(df_ts) > 0:
-                _df = pd.concat(df_ts).drop_duplicates()
-            return _df
-
-        def _use_model():
-            return self.context.run_model('account.token-transfer',
-                                          input={'address': lp, 'tokens': [pool.address]},
-                                          return_type=Records).to_dataframe()
-
-        # _df = _use_model().rename(columns={'value': 'transaction_value'})[q_cols]
-        # _df.loc[1:, 'transaction_value'] = -1 * (_df.loc[0, 'transaction_value'])
-
-        def _use_events():
-            minted = pd.DataFrame(pool.fetch_events(pool.events.Transfer, argument_filters={
-                'to': lp.checksum}, from_block=0, to_block=self.context.block_number,
-                contract_address=pool.address.checksum))
-            burnt = pd.DataFrame(pool.fetch_events(pool.events.Transfer, argument_filters={
-                'from': lp.checksum}, from_block=0, to_block=self.context.block_number,
-                contract_address=pool.address.checksum))
-            df_empty = pd.DataFrame(data=[], columns=['transactionHash',
-                                    'blockNumber', 'logIndex', 'from', 'to', 'value'])
-            df_combined = (pd.concat(
-                [minted.loc[:, ['transactionHash', 'blockNumber', 'logIndex', 'from', 'to', 'value']] if not minted.empty else df_empty,
-                 (burnt.loc[:, ['transactionHash', 'blockNumber', 'logIndex', 'from', 'to', 'value']].assign(
-                     value=lambda x: -x.value) if not burnt.empty else df_empty)
-                 ])
-                .assign(transactionHash=lambda r: r.transactionHash.apply(lambda x: x.hex()))
-                .sort_values(['blockNumber', 'logIndex'])
-                .rename(columns={
-                    'transactionHash': 'transaction_hash',
-                    'blockNumber': 'block_number',
-                    'logIndex': 'log_index',
-                    'from': 'from_address',
-                    'to': 'to_address',
-                    'value': 'transaction_value'})
-                .reset_index(drop=True))
-
-            return df_combined
-
-        _df = _use_events()
-
-        if _df.empty:
-            return Records.from_dataframe(_df)
-
-        if (_df["block_number"].tail(1) != int(self.context.block_number)).all():
-            new_row = [('', int(self.context.block_number), -1, input.lp, input.lp, 0)]
-            _df = pd.concat([_df, pd.DataFrame(new_row, columns=q_cols)]).reset_index(drop=True)
-
-        lp_prev_token0 = 0
-        lp_prev_token1 = 0
-
-        for row_n, row_data in _df.iterrows():
-            # Uniswap V2 has compounding effect for the fee
-            # fee = current lp position - IL'ed position from previous lp position - current addition/removal
-            block_number = row_data['block_number']
-            transaction_value = row_data['transaction_value']
-
-            v2_fee = calculate_v2_fee(
-                self.context, pool, lp, block_number, transaction_value,
-                lp_prev_token0, lp_prev_token1)
-
-            lp_prev_token0 = v2_fee['token0_lp']
-            lp_prev_token1 = v2_fee['token1_lp']
-
-            for it in ['token0_lp', 'token1_lp',
-                       'in_out_amount0', 'in_out_amount1',
-                       'token0_lp_current', 'token1_lp_current',
-                       'token0_fee', 'token1_fee']:
-                _df.loc[row_n, it] = try_zero(v2_fee[it])  # type: ignore
-
-        return Records.from_dataframe(_df)
-
-
-@Model.describe(slug='uniswap-v2.lp-fee',
-                version='0.7',
-                display_name='Uniswap v2 (Sushiswap) LP Position (split for fee) for account',
-                description='Returns position (split for fee) for account. The fee is accumulated from last position change.',
-                category='protocol',
-                subcategory='uniswap-v2',
-                input=V2LPInput,
-                output=V2LPFeeOutput)
-class UniswapV2LPFee(Model):
-    def run(self, input: V2LPInput) -> V2LPFeeOutput:
-        pool = input.pool
-        lp = input.lp
-
-        try:
-            _ = pool.abi
-        except ModelDataError:
-            pool.set_abi(UNISWAP_V2_POOL_ABI)
-
-        token0 = Token(address=Address(pool.functions.token0().call()))
-        token1 = Token(address=Address(pool.functions.token1().call()))
-        token0 = token0.as_erc20(set_loaded=True)
-        token1 = token1.as_erc20(set_loaded=True)
-
-        # Obtain the last 2 when the current block has mint/burn
-        with self.context.ledger.TokenBalance as q:
-            _df = q.select(
-                aggregates=[(q.TRANSACTION_VALUE.sum_(), 'sum_transaction_value')],
-                order_by=q.BLOCK_NUMBER.desc(),
-                where=(q.ADDRESS.eq(lp)
-                        .and_(q.TOKEN_ADDRESS.eq(pool.address))
-                       ),
-                group_by=[q.BLOCK_NUMBER],
-                limit=2
-            ).to_dataframe()
-
-        if _df.empty:
-            return V2LPFeeOutput.zero(lp, token0, token1)
-
-        prev_block_number = _df['block_number'].to_list()[0]
-        prev_transaction_value = _df['sum_transaction_value'].to_list()[0]
-
-        if prev_block_number == self.context.block_number:
-            prev2_block_number = _df['block_number'].to_list()[1]
-            lp_pos = self.context.run_model(
-                'uniswap-v2.lp',
-                input=V2LPInput(pool=input.pool, lp=input.lp),
-                return_type=V2LPOutput,
-                block_number=prev2_block_number)
-
-            lp_prev_token0, lp_prev_token1 = lp_pos.tokens[0].amount, lp_pos.tokens[1].amount
-
-            v2_fee = calculate_v2_fee(
-                self.context, pool, lp, prev_block_number, prev_transaction_value,
-                lp_prev_token0, lp_prev_token1)
-        else:
-            lp_pos = self.context.run_model(
-                'uniswap-v2.lp',
-                input=V2LPInput(pool=input.pool, lp=input.lp),
-                return_type=V2LPOutput,
-                block_number=prev_block_number)
-
-            lp_prev_token0, lp_prev_token1 = lp_pos.tokens[0].amount, lp_pos.tokens[1].amount
-
-            v2_fee = calculate_v2_fee(
-                self.context, pool, lp, self.context.block_number, 0,
-                lp_prev_token0, lp_prev_token1)
-
-        lp_balance = pool.functions.balanceOf(lp).call()
-
-        lp_position = Position(amount=pool.scaled(lp_balance), asset=pool)
-        position0 = PositionWithFee(
-            amount=v2_fee['token0_lp'] - v2_fee['token0_fee'],
-            fee=v2_fee['token0_fee'],
-            asset=token0)
-        position1 = PositionWithFee(
-            amount=v2_fee['token1_lp'] - v2_fee['token1_fee'],
-            fee=v2_fee['token1_fee'],
-            asset=token1)
-
-        return V2LPFeeOutput(lp=lp_position,
-                             tokens=[position0, position1])
+        factory_addr = V2_FACTORY_ADDRESS[self.context.network]
+        return self.get_uniswap_pools_ledger(self.context, factory_addr, input.address)
 
 
 @Model.describe(slug='uniswap-v2.get-pool-price-info',
-                version='1.12',
+                version='1.13',
                 display_name='Uniswap v2 Token Pool Price Info',
                 description='Gather price and liquidity information from pool',
                 category='protocol',
@@ -528,10 +236,10 @@ class UniswapPoolPriceInfo(Model):
             pool = Contract(address=input.address)
             pool.set_abi(abi=UNISWAP_V2_POOL_ABI, set_loaded=True)
 
-        primary_tokens = self.context.run_model('dex.ring0-tokens',
-                                                input=EmptyInput(),
-                                                return_type=Some[Address],
-                                                local=True).some
+        ring0_tokens = self.context.run_model('dex.ring0-tokens',
+                                              input=EmptyInput(),
+                                              return_type=Some[Address],
+                                              local=True).some
         weth_address = Token('WETH').address
 
         reserves = pool.functions.getReserves().call()
@@ -580,11 +288,18 @@ class UniswapPoolPriceInfo(Model):
         # 3. If WETH-WBTC: use WETH to price WBTC
         # 4. If SB/WETH/WBTC-X: use SB/WETH/WBTC to price
 
-        is_primary_pool = token0.address in primary_tokens and token1.address in primary_tokens
+        is_ring0_pool = token0.address in ring0_tokens and token1.address in ring0_tokens
 
-        if is_primary_pool:
-            pass
-        elif token0.address == weth_address and token1.address in primary_tokens:
+        if is_ring0_pool:
+            ref_price_ring0 = self.context.run_model(
+                slug=input.ref_price_slug,
+                input=EmptyInput(),
+                return_type=dict)
+            # Use the reference price to scale the tick price. Note the cross-reference is used here.
+            # token0 = tick_price0 * token1 = tick_price0 * ref_price of token1
+            tick_price0 *= ref_price_ring0[token1.address]
+            tick_price1 *= ref_price_ring0[token0.address]
+        elif token0.address == weth_address and token1.address in ring0_tokens:
             ref_price = self.context.run_model(
                 slug=input.price_slug,
                 input=DexPriceTokenInput(
@@ -593,7 +308,7 @@ class UniswapPoolPriceInfo(Model):
                     debug=input.debug),
                 return_type=Price,
                 local=True).price
-        elif token1.address == weth_address and token0.address in primary_tokens:
+        elif token1.address == weth_address and token0.address in ring0_tokens:
             ref_price = self.context.run_model(
                 slug=input.price_slug,
                 input=DexPriceTokenInput(
@@ -607,9 +322,9 @@ class UniswapPoolPriceInfo(Model):
         # elif token1.address == weth_address and token0.address == wbtc_address:
         # use token1 as reference token
         else:
-            if token0.address in primary_tokens + [weth_address]:
+            if token0.address in ring0_tokens + [weth_address]:
                 primary_address = token0.address
-            elif token1.address in primary_tokens + [weth_address]:
+            elif token1.address in ring0_tokens + [weth_address]:
                 primary_address = token1.address
             else:
                 primary_address = Address.null()
@@ -627,8 +342,9 @@ class UniswapPoolPriceInfo(Model):
                     raise ModelRunError(f'Can not retrieve price for '
                                         f'{Token(address=primary_address)}')
             else:
-                self.logger.warning('There is no primary token in this pool: '
-                                    f'{token0.address}/{token0.symbol} and {token1.address}/{token1.symbol}')
+                self.logger.warning(
+                    'There is no primary token in this pool: '
+                    f'{token0.address}/{token0.symbol} and {token1.address}/{token1.symbol}')
                 ref_price = 0
 
         pool_price_info = PoolPriceInfo(src=input.price_slug,
@@ -658,17 +374,16 @@ class UniswapPoolPriceInfo(Model):
                 input=DexPriceTokenInput,
                 output=Some[PoolPriceInfo])
 class UniswapV2GetTokenPriceInfo(Model):
-    def run(self, input: DexPriceTokenInput) -> Some[PoolPriceInfo]:
-        pools = self.context.run_model('uniswap-v2.get-pools',
-                                       input,
-                                       return_type=Contracts,
-                                       local=True)
-
+    def get_pools_info(
+            self,
+            input: DexPriceTokenInput,
+            pools: Contracts) -> Some[PoolPriceInfo]:
         model_slug = 'uniswap-v2.get-pool-price-info'
         model_inputs = [
             DexPricePoolInput(
                 address=pool.address,
                 price_slug='uniswap-v2.get-weighted-price',
+                ref_price_slug='uniswap-v2.get-ring0-ref-price',
                 weight_power=input.weight_power,
                 debug=input.debug)
             for pool in pools]
@@ -707,8 +422,15 @@ class UniswapV2GetTokenPriceInfo(Model):
             return infos
 
         infos = _use_for(local=True)
-
         return Some[PoolPriceInfo](some=infos)
+
+    def run(self, input: DexPriceTokenInput) -> Some[PoolPriceInfo]:
+        pools = self.context.run_model('uniswap-v2.get-pools',
+                                       input,
+                                       return_type=Contracts,
+                                       local=True)
+
+        return self.get_pools_info(input, pools)
 
 
 class UniswapV2PoolInfo(DTO):
